+++ src/commands/show-quick-pick.ts
@@ -1,16 +1,32 @@
 import { commands, QuickPickItem, window } from 'vscode'
 import { contributes } from '../../package.json'
+import { command as exportCommand } from './export'
+import { command as toggleMarpPreviewCommand } from './toggle-marp-preview'
+import { command as openExtensionSettingsCommand } from './open-extension-settings'
 
-const availableCommands: QuickPickItem[] = []
+export const cmdSymbol = Symbol()
 
-for (const command of contributes.commands) {
-  if (command.command === 'markdown.marp.showQuickPick') continue
+const contributedCommands = [exportCommand, toggleMarpPreviewCommand]
+const availableCommands: (QuickPickItem & { [cmdSymbol]: string })[] = []
 
-  availableCommands.push({
-    label: command.title,
-    description: command.command,
-  })
+for (const cmdPath of contributedCommands) {
+  const cmd = contributes.commands.find(({ command }) => cmdPath === command)
+
+  if (cmd) {
+    availableCommands.push({
+      [cmdSymbol]: cmdPath,
+      description: cmdPath,
+      label: cmd.title,
+    })
+  }
 }
+
+availableCommands.push({
+  label: '$(settings-gear) Open extension settings',
+  [cmdSymbol]: openExtensionSettingsCommand,
+})
+
+export const command = 'markdown.marp.showQuickPick'
 
 export default async function showQuickPick() {
   const command = await window.showQuickPick(availableCommands, {
@@ -18,5 +34,7 @@ export default async function showQuickPick() {
     placeHolder: 'Select available command in Marp for VS Code...',
   })
 
-  if (command?.description) await commands.executeCommand(command.description)
+  if (command?.[cmdSymbol]) {
+    await commands.executeCommand(command[cmdSymbol])
+  }
 }